Add CDTDatastore tests to build script
Added ability to run tests from the CDTDatastore workspace, in addition
to replication acceptance tests.

ReplicationAcceptance tests are only run when -nightly true is passed
into the build script. if -nightly false is passed in or nothing is
passed in the build script will only run the tests from CDTDatastore